requirements.txt
apache-airflow
pandas
snowflake-connector-python
tableauserverclient
flask
pytest
scipy
